[#4] Accept code.xml and .catrobat file extension
All programs must have a /program/header/url set, which only the official
share server will do. As workaround you can

- Set the /program/header/url to a valid url manually.
- Use the code.xml file and rename it following the format
  MILLISECONDS_ID.xml, e.g. 1364288279000_710.xml.

Use a unique id for each program, otherwise they may not show up in the
statistics.

The upload date for code.xml and .catrobat files can't be determined, so
new Date(0) is used. Such programs will appear only in the overall
statistics, not in the monthly or yearly ones.